Major improvement in standardization of projects, this project succeeds the following checks done by the "AnalyseProjects" app: Versioned, Folder structures, Filepaths, MainWin/form, AppIcon, AutoUpdating, AboutBox. Licensing not implemented and other checks must still be programmed to be done (in AnalyseProjects).
This app now works great, can drag-drop files/folders onto it to add and right-click menu to remove. It makes use of Windows 7 JumpLists on taskbar.

Right-click menu of the window (not one of the shorcuts) currently has two "Create from Clipboard" options.
"App from text": makes use of the current text (if any) in the Clipboard, and parses the text to see if it can find a file/directory with/without commandline-arguments.
"Chrome app from Url": Takes the Url from the Clipboard (if there is a valid Url) and creates a shortcut to make it launch as a "Google Chrome Application".

git-svn-id: svn://localhost/taskbarshortcuts/trunk@3 871e9758-bc6e-2443-8d1e-c268101c5431